move event pool class to separate file
